Improve class structure
