+++ src/main/java/edu/ie3/datamodel/io/csv/FileNameMetaInformation.java
@@ -1,0 +1,9 @@
+/*
+ * © 2020. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+*/
+package edu.ie3.datamodel.io.csv;
+
+/** Meta information, that can be derived from a certain file name */
+public interface FileNameMetaInformation {}
