+++ src/main/java/edu/ie3/datamodel/io/csv/FileNamingStrategy.java
@@ -5,6 +5,9 @@
 */
 package edu.ie3.datamodel.io.csv;
 
+import edu.ie3.datamodel.io.csv.timeseries.ColumnScheme;
+import edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation;
+import edu.ie3.datamodel.io.csv.timeseries.LoadProfileTimeSeriesMetaInformation;
 import edu.ie3.datamodel.models.UniqueEntity;
 import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.AssetTypeInput;
@@ -21,8 +24,6 @@ import edu.ie3.datamodel.models.timeseries.repetitive.LoadProfileInput;
 import edu.ie3.datamodel.models.value.*;
 import edu.ie3.util.StringUtils;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.regex.Matcher;
@@ -181,8 +182,7 @@ public class FileNamingStrategy {
       Optional<E> maybeFirstElement = timeSeries.getEntries().stream().findFirst();
       if (maybeFirstElement.isPresent()) {
         Class<? extends Value> valueClass = maybeFirstElement.get().getValue().getClass();
-        Optional<IndividualTimeSeriesMetaInformation.ColumnScheme> mayBeColumnScheme =
-            IndividualTimeSeriesMetaInformation.ColumnScheme.parse(valueClass);
+        Optional<ColumnScheme> mayBeColumnScheme = ColumnScheme.parse(valueClass);
         if (mayBeColumnScheme.isPresent()) {
           return Optional.of(
               prefix
@@ -263,8 +263,8 @@ public class FileNamingStrategy {
           "Cannot extract meta information on individual time series from '" + fileName + "'.");
 
     String columnSchemeKey = matcher.group("columnScheme");
-    IndividualTimeSeriesMetaInformation.ColumnScheme columnScheme =
-        IndividualTimeSeriesMetaInformation.ColumnScheme.parse(columnSchemeKey)
+    ColumnScheme columnScheme =
+        ColumnScheme.parse(columnSchemeKey)
             .orElseThrow(
                 () ->
                     new IllegalArgumentException(
@@ -407,112 +407,5 @@ public class FileNamingStrategy {
    */
   private String addPrefixAndSuffix(String s) {
     return prefix.concat(s).concat(suffix);
-  }
-
-  public interface FileNameMetaInformation {}
-
-  public static class IndividualTimeSeriesMetaInformation implements FileNameMetaInformation {
-    public enum ColumnScheme {
-      ENERGY_PRICE("c"),
-      ACTIVE_POWER("p"),
-      APPARENT_POWER("pq"),
-      HEAT_DEMAND("h"),
-      ACTIVE_POWER_AND_HEAT_DEMAND("ph"),
-      APPARENT_POWER_AND_HEAT_DEMAND("pqh"),
-      WEATHER("weather");
-
-      private final String scheme;
-
-      ColumnScheme(String scheme) {
-        this.scheme = scheme;
-      }
-
-      public String getScheme() {
-        return scheme;
-      }
-
-      public static Optional<ColumnScheme> parse(String key) {
-        String cleanString = StringUtils.cleanString(key).toLowerCase();
-        return Arrays.stream(ColumnScheme.values())
-            .filter(entry -> Objects.equals(entry.scheme, cleanString))
-            .findFirst();
-      }
-
-      public static <V extends Value> Optional<ColumnScheme> parse(Class<V> valueClass) {
-        /* IMPORTANT NOTE: Make sure to start with child classes and then use parent classes to allow for most precise
-         * parsing (child class instances are also assignable to parent classes) */
-
-        if (EnergyPriceValue.class.isAssignableFrom(valueClass)) return Optional.of(ENERGY_PRICE);
-        if (HeatAndSValue.class.isAssignableFrom(valueClass))
-          return Optional.of(APPARENT_POWER_AND_HEAT_DEMAND);
-        if (SValue.class.isAssignableFrom(valueClass)) return Optional.of(APPARENT_POWER);
-        if (HeatAndPValue.class.isAssignableFrom(valueClass))
-          return Optional.of(ACTIVE_POWER_AND_HEAT_DEMAND);
-        if (PValue.class.isAssignableFrom(valueClass)) return Optional.of(ACTIVE_POWER);
-        if (HeatDemandValue.class.isAssignableFrom(valueClass)) return Optional.of(HEAT_DEMAND);
-        if (WeatherValue.class.isAssignableFrom(valueClass)) return Optional.of(WEATHER);
-        return Optional.empty();
-      }
-    }
-
-    private final UUID uuid;
-    private final ColumnScheme columnScheme;
-
-    public IndividualTimeSeriesMetaInformation(UUID uuid, ColumnScheme columnScheme) {
-      this.uuid = uuid;
-      this.columnScheme = columnScheme;
-    }
-
-    public UUID getUuid() {
-      return uuid;
-    }
-
-    public ColumnScheme getColumnScheme() {
-      return columnScheme;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (!(o instanceof IndividualTimeSeriesMetaInformation)) return false;
-      IndividualTimeSeriesMetaInformation that = (IndividualTimeSeriesMetaInformation) o;
-      return uuid.equals(that.uuid) && columnScheme == that.columnScheme;
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(uuid, columnScheme);
-    }
-  }
-
-  public static class LoadProfileTimeSeriesMetaInformation implements FileNameMetaInformation {
-    private final UUID uuid;
-    private final String profile;
-
-    public LoadProfileTimeSeriesMetaInformation(UUID uuid, String profile) {
-      this.uuid = uuid;
-      this.profile = profile;
-    }
-
-    public UUID getUuid() {
-      return uuid;
-    }
-
-    public String getProfile() {
-      return profile;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (!(o instanceof LoadProfileTimeSeriesMetaInformation)) return false;
-      LoadProfileTimeSeriesMetaInformation that = (LoadProfileTimeSeriesMetaInformation) o;
-      return uuid.equals(that.uuid) && profile.equals(that.profile);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(uuid, profile);
-    }
   }
 }
